Add test folder and pronect source folder
Add a folder for testing and also a folder for the source files

The test folder will have a unit and integration test seperate folders.

The src folder follows the concept of layered architecture.
The following sub-folders are created in the src folder:
-models : This folder contains schema definition for our documents

-controller: The business logic of our application will reside in This
folder

-lib : Helper functions such as parsing of date are co-placed in this
folder

-services: The services folder serves as the folder for Data Access
Layer codes

- routes: The endpoints for communicating with the services

- middleware: Functions that operate on req object are placed here

- loader :  Configurations for MongoDB connections , redis launc and
connection to ElasticSearch are placed here