Test add question method in question service
Test the method handling adding of question to ensure the method works.
This should be done through the use of sinon and its stub method.

Seperate the network file from the app declaration by creating a bin/www
path at the root directory.

Add a controller for question and also add initial code logic into its

Make export of functionality from folders reside in a single index file
at the root of the folder of concern.
This changes were applied to the services and models aspect of the project

Add a new but empty file for unit testing of middleware and also
for testing of the controller codes