Improve question cache service
The cache service was using a hash for storing retrieve questions.

This was discovered not to be suitable has more than one key has to be
defined to be able to retrieve a question from redis.

The structure for a question was converted to a string.
The question is stringnify first by using JSON.stringnify(question)
before saving.

Other redis client method that were earlier promisified were dropped
as they add no use in our current application